Attempting to switch the scores from floats in 0.-1. to ints in 0-100.
YOU MAY NEED TO CHANGE YOUR FILTER RULES ACCORDINGLY.

The Rule dialog appears to have gotten partly broken before this --
referenced a _GetFolder method that doesn't exist.  Repaired that.

PROBLEM:  While the Rule dialog seems to work fine now, trying to write
a score to the SpamProb (Hammie, whatever) custom field doesn't appear
to do anything anymore.  But I didn't touch that part of the code, so
I'm baffled.  Perhaps it has to do with that the *type* changed from
double to int, and that _MapiTypeMap can't tell the difference between
an int and a bool before Python 2.3 (I'm using 2.2)?  But commenting
out the PT_BOOLEAN entry there didn't make any difference for me, so I
doubt that's it.