Fix a couple of mistakes on the apps page, plus update the information about the binary's requirements and known problems.
Update developer page for new status of Python 2.3 and
current date.

Add plural of corpus to glossary on docs page.

Cosmetic fix to download page.

Added notes (based on Tim P's post) about using the outlook plugin
to the windows page.

Fixed the script so that the sourceforge logo links to sf, not
to us.