2 changes to the way the DB classifier manages words:
* As per Tim P's mail, keep a list of "changed words" with a flag
indicating "change" or "delete".  This prevents the database save
from updating every single word ever loaded by the db.

* From Sean, a change that prevents caching of hapaxes.  Such words are
saved directly to the DB.  This reduces the memory footprint significantly
(as these words are not kept in memory) and helps save times.

This change makes "incremental" saving of the database happen in a
reasonable time, and doesn't degrade after a complete retrain etc.

I'm off for a weekend holiday - someone can just back this out if I
screwed it up <wink>