Fixed a bug whereby long attachments would be broken by the POP3 proxy.
For messages that take a long time to download, the proxy starts
to forward them after 30 seconds, regardless of whether it's seen the
whole message.  It classifies the message using however much of the
message it's seen so far.  It's done this since the early days, but in the
intervening time someone (whose anonimiTy I MuST hONour hEre)
has changed it to parse the message and then use the stringified
version of the parsed message to forward to the client.  The problem is
that the parser will 'fix' a partial message by appending a closing
boundary separator - this separator can then end up embedded in the
middle of a long attachment.

We still parse the message and add the Spambayes headers to the
parsed message, but we now take just the headers from the parsed
message and add the (possible partial) body ourselves.

This edit simplifies Scott Schlesier's trailing dot fix, because we no
longer need to cope with added boundary separators.