Little tool to find messages with duplicate property values.  Useful to find message that SpamBayes will consider duplicate.
Example, to messages with the same PR_SEARCH_KEY property in 2 folders:

% find_dupe_props.py -f "\Good spam-looing mail" -f Inbox PR_SEARCH_KEY
Folder '... mail': 48 items with the property and 0 items without it
Folder 'Inbox': 1699 items with the property and 0 items without it
Found 2 items with property value '...fK\xb0\x95\xb9\xa7s\xce\x93\xcd'
 Courtesy notification - 50% of your current month's allowance used
 Courtesy notification - 50% of your current month's allowance used
...

This output shows the folders scanned, and the subjects for all messages
with an identical PR_SEARCH_KEY property (which is the property used by
SpamBayes to track training data).  This shows that 2 items with identical
subjects have the same ID - ie, these are copies of the same message.