Few tweaks to the logging code (Outlook UI events are reported at level 3, and more useful logging of the timers.)
The timer now relies also on "HaveSeenMessage()" along with the msgstore
generator.  The msgstore generator assumes that we have saved spam scores,
which meant that if we didn't save scores, we would continually re-try the
same old unread messages each time a new one arrived.  Now, if these msgs
have been trained on, they are skipped, meaning we get to the new messages
quicker.