A bug report had 'text' ending up as None, which stoped the messages displaying.
This handles that error, so that things keep working, at least, but it would be good
to figure out how it is that 'text' can become None.  As far as I can see, it means
that get_payload() is returning None, but I don't think it's meant to be able to do that.