Stupid global variables!
Thanks to a global variable not being updated, when we recreated everything, the userinterface kept using the old classifier.
Since we now behave and close that one, this caused all sorts of problems.  Get rid of the damn global variable, and correctly
update it, and all is well in the world again.

In addition, don't save an empty database.  I think we make assumptions about the db being non-empty in some places.

This should fix [ spambayes-Bugs-809769 ] TypeError when training 1.0a6

(I can't believe it took so long for me to find this!)