Fix [ 814322 ] AttributeError: hdrtxt
Basically the review page would die if one of the cached messages was moved from
the cache directory by someone other than spambayes (for example a virus
protection program).  Instead, we ignore the problem and just don't present that message
for review.

Note that you can actually still train some messages that have been removed,
as long as they are in the memory cache.  I can't think of any problems that this
would cause.