Two new FAQs:
Missing mail: it's odd that this is very suddenly a FAQ, and nowhere near a release.
I'm confident that it's not the plug-in, though, so that's what I've said.  I've put some
general 'how to find mail' tips there to try and help people along.
(Anyone else wondering if ISPs are to blame because they've put in too stringent
filters to try and block all the worm crud recently?)

Uninstall:  how to uninstall the plug-in, including removing the toolbar (the oldest
still-open bug!).  I'm pretty sure that this is the current situation...