Improve the messageinfo database.  Previously it was fixed to store two attributes (classification and training), but the idea was that it was meant to be expandable so that other attributes can be saved (for example, sb_pop3dnd saves flags, and Anthony wants to save message-ids).  As it was, this meant that the db's couldn't be backwards compatible, which isn't nice.
The scheme now works for old messageinfo dbs (so nothing should break),
but will let you save any number of different attributes, simply by changing
an attribute of the message object (which means that different messages
can even save different things, if that's what is wanted).

Also use the relative-to-ini not relative-to-cwd pathnames stuff with the options.