For any regex defined options, allow a blank string (i.e. no answer) to be a valid response. I'm sure this was already the case, but it doesn't seem so.
Add a third configuration page.  This will automatically include all current experimental
(but not deprecated, as long as we all stuck to putting (DEPRECATED) at the start of
the description) options.  Maybe people will test things for us.  The link to this is on the
main configuration page; maybe it should be on the advanced page?