The earlier change to using BODY.PEEK was faulty in two ways, so fix that.
 1. Python's imaplib is quirky in that it quotes when it shouldn't, unless you put
the command in (), so do that.

 2. The extract_fetch_data function wasn't setup to handle BODY.PEEK[]
responses, and then the creation of the message from that wasn't set up to
handle them either, so add those in.

My testing indicates that (at least with my server) using BODY.PEEK[] now works,
and retains the /Seen status as it should.

(Note that even during the failing, it would still work, since it would fall back to
RFC822, but it would lose the /Seen status).