Add a comment to mboxutils.
mboxtrain is the last (I think) script that still gets hit by the TypeError problem, so fix that.
For the moment, this fix is just to apply the fix that Skip made for sb_filter - i.e. instead
of calling msg.as_string(), we call mboxutils.as_string(msg).

I can't see how this would break anything, unless the mboxutils.as_string code is broken,
which is already in 1.0, so this seems to be a good bugfix candidate.

Closes [ 942262 ] TypeError: string payload expected: <type 'list'>