2.3 compatibility: add reversed() function
When deciding if messages are misses or not, consider whether they have
Message-Id or Subject headers.  If they have neither, there's probably
something funky going on.  In that case, just ignore them altogether.
Maybe it's not the best approach, but it's the best I've come up with so far
until I figure out what the heck's going on.