Add a "where is my recover from spam button" FAQ.
No-one seems to answer me saying "oh yes, that was it" when I give this
answer, but it should be right, and putting it here saves me typing it again
and again and again.  If anyone can think of any other reasons the button
might not appear, please add them!