As per the discussion on spambayes-dev.
Rather than treating the notate_to option just like notate_subject,
we instead convert the classification into an email address in the form
(eg) spam@spambayes.invalid and adds that as a separate address to
the message.

This has the added benefit of being a legitimate treatment of the header,
and also should be easier to use in a message rule, because there should
be no other mail at all that would match a rule looking for @spambayes.invalid.

If the message is replied to (say it was unsure ham) then the user can either
delete the invalid address, or leave it and get a bounce back (but the legit
address will still get the reply).